Python/src/requirements.txt
semantic-kernel
asyncio
python-dotenv
azure-search-documents
fastapi
pandas
uvicorn
streamlit
pydantic
autogen-agentchat
autogen-ext[openai,azure]